reorganized structure to avoid circular imports
