+++ theme/icons_test.go
@@ -4,16 +4,26 @@ import (
 	"fyne.io/fyne"
 	_ "fyne.io/fyne/test"
 	"github.com/stretchr/testify/assert"
+	"io/ioutil"
+	"path/filepath"
 	"testing"
 )
 
-// TODO: add tests back in
 func helperNewStaticResource() *fyne.StaticResource {
 	return &fyne.StaticResource{
 		StaticName: "content-remove.svg",
 		StaticContent: []byte{
 			60, 115, 118, 103, 32, 120, 109, 108, 110, 115, 61, 34, 104, 116, 116, 112, 58, 47, 47, 119, 119, 119, 46, 119, 51, 46, 111, 114, 103, 47, 50, 48, 48, 48, 47, 115, 118, 103, 34, 32, 119, 105, 100, 116, 104, 61, 34, 50, 52, 34, 32, 104, 101, 105, 103, 104, 116, 61, 34, 50, 52, 34, 32, 118, 105, 101, 119, 66, 111, 120, 61, 34, 48, 32, 48, 32, 50, 52, 32, 50, 52, 34, 62, 60, 112, 97, 116, 104, 32, 102, 105, 108, 108, 61, 34, 35, 102, 102, 102, 102, 102, 102, 34, 32, 100, 61, 34, 77, 49, 57, 32, 49, 51, 72, 53, 118, 45, 50, 104, 49, 52, 118, 50, 122, 34, 47, 62, 60, 112, 97, 116, 104, 32, 100, 61, 34, 77, 48, 32, 48, 104, 50, 52, 118, 50, 52, 72, 48, 122, 34, 32, 102, 105, 108, 108, 61, 34, 110, 111, 110, 101, 34, 47, 62, 60, 47, 115, 118, 103, 62},
 	}
+}
+
+func helperLoadBytes(t *testing.T, name string) []byte {
+	path := filepath.Join("testdata", name) // path relative to this file
+	bytes, err := ioutil.ReadFile(path)
+	if err != nil {
+		t.Fatal(err)
+	}
+	return bytes
 }
 
 func TestIconThemeChangeDoesNotChangeName(t *testing.T) {
@@ -56,6 +66,56 @@ func TestNewThemedResource_OneStaticResourceSupport(t *testing.T) {
 	assert.Equal(t, nm, custom.Name())
 }
 
+func TestDynamicResource_Name(t *testing.T) {
+	sr := fyne.NewStaticResource("cancel_Paths.svg",
+		helperLoadBytes(t, "cancel_Paths.svg"))
+	dr := &DynamicResource{
+		BaseResource: sr,
+	}
+	assert.Equal(t, sr.Name(), dr.Name())
+}
+
+func TestDynamicResource_Content_NoGroupsFile(t *testing.T) {
+	fyne.CurrentApp().Settings().SetTheme(DarkTheme())
+	sr := fyne.NewStaticResource("cancel_Paths.svg",
+		helperLoadBytes(t, "cancel_Paths.svg"))
+	dr := &DynamicResource{
+		BaseResource: sr,
+	}
+	assert.NotEqual(t, sr.Content(), dr.Content())
+}
+
+func TestDynamicResource_Content_GroupPathFile(t *testing.T) {
+	fyne.CurrentApp().Settings().SetTheme(DarkTheme())
+	sr := fyne.NewStaticResource("check_GroupPaths.svg",
+		helperLoadBytes(t, "check_GroupPaths.svg"))
+	dr := &DynamicResource{
+		BaseResource: sr,
+	}
+	assert.NotEqual(t, sr.Content(), dr.Content())
+}
+
+func TestDynamicResource_Content_GroupRectFile(t *testing.T) {
+	fyne.CurrentApp().Settings().SetTheme(DarkTheme())
+	sr := fyne.NewStaticResource("info_GroupRects.svg",
+		helperLoadBytes(t, "info_GroupRects.svg"))
+	dr := &DynamicResource{
+		BaseResource: sr,
+	}
+	assert.NotEqual(t, sr.Content(), dr.Content())
+}
+
+func TestDynamicResource_Content_GroupPolygonsFile(t *testing.T) {
+	fyne.CurrentApp().Settings().SetTheme(DarkTheme())
+	sr := fyne.NewStaticResource("warning_GroupPolygons.svg",
+		helperLoadBytes(t, "warning_GroupPolygons.svg"))
+	dr := &DynamicResource{
+		BaseResource: sr,
+	}
+	assert.NotEqual(t, sr.Content(), dr.Content())
+}
+
+// Test Asset Sources
 func Test_FyneLogo(t *testing.T) {
 	fyne.CurrentApp().Settings().SetTheme(DarkTheme())
 	result := FyneLogo().Name()
